Add files via upload
Cylindrical maps of Jupiter 5-µm images captured between 1999-2000 and 2006-2007 by NSFCam and NSFCam2 instruments, respectively, mounted on the Infrared Telescope Facility (IRTF) on Maunakea. As explained in Antuñano et al. 2018, the data was reduced using the Data Reduction Manager (DRM) (Fletcher et al., 2009) and was projected in cylindrical maps of meridional and zonal resolutions of 0.5°-1°.